move utils to internal
