+++ core/config/config.go
@@ -3,7 +3,7 @@ package config
 import (
 	"github.com/invopop/jsonschema"
 	"github.com/railwayapp/railpack/core/plan"
-	"github.com/railwayapp/railpack/core/utils"
+	"github.com/railwayapp/railpack/internal/utils"
 )
 
 const (
